mish: use process groups to kill sh runs (#1905)
* mish: use process groups to kill sh runs

If your `sh` starts a process, that starts child processes then canceling
the context will _not_ cause the child process to be killed. This is a
problem if you are starting processes that are servers because the dangling
process will still be using the port that the new server wants to use.

To get around this we start processes with PGID. Then, we use a special
form of Signal to kill all of the processes in the process group.

If someone wrote a server/process spawner that changes the process group
on start we would be out of luck. :(

* Fix case where goroutine would never finish

GitOrigin-RevId: d2905e785d9225a9ddcf0c0902201a9c1001b475